\name{.q}
\alias{.q}
\title{Quit the R session without saving your workspace.}
\usage{.q(...)}
\description{Quit the R session without saving your workspace.}
\details{The save workspace[y/n] prompt is fucking annoying. Avoid it with .q( ) or qn( )}
\note{Part of a suite of Opower packages}
\author{Erik Shilts - Shamelessly ripped from stackoverflow}
\arguments{\item{...}{other parameters for q()}}
