\name{BottomCode}
\alias{BottomCode}
<<<<<<< HEAD
\title{Set a Minimum Value For a Numeric Vector}
=======
\title{Set a minimum value for a numeric vector}
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\usage{
  BottomCode(vec, level = 0)
}
\arguments{
  \item{vec}{a numeric vector}

<<<<<<< HEAD
  \item{level}{the minimum value; default is 0}
}
\value{
  a vector with all values < \code{level} set to
  \code{level}
=======
  \item{level}{the minimum value. Default is 0.}
}
\value{
  a vector with all values < level set to level
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
}
\description{
  Bottom codes a vector by setting a minimum value
}
\note{
  Part of a suite of Opower packages
}
<<<<<<< HEAD
\examples{
BottomCode(1:10, level=5)
BottomCode(1:10, level=11)
BottomCode(1:10, level=0)
}
=======
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\author{
  Erik Shilts
}

