\name{classes}
\alias{classes}
\title{Classes of Columns in a Dataframe}
\usage{
  classes(dat)
}
\arguments{
  \item{dat}{a data.frame}
}
\value{
  a vector of class names
}
\description{
  Equivalent to \code{cbind(sapply(dataframe, class))}
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts
}

