\name{concat}
\alias{concat}
\title{A way to paste strings together without having to set sep = fucking '' over and over again.}
\usage{
  concat(...)
}
\arguments{
  \item{...}{strings to concatenate}
}
\value{
  a concatenated string. meow.
}
\description{
  Seriously, this should be obvious.
}
\note{
  Part of a suite of Opower packages
}
\author{
  Yoni Ben-Smellshulam
}
\seealso{
  \code{\link{paste}}
}

