\name{DFInfo}
\alias{DFInfo}
\title{Useful Information on a Given Dataframe
Returns a dataframe with the names of each of the columns, the class, and number of missing observations}
\usage{
  DFInfo(dat)
}
\arguments{
  \item{dat}{a data frame}
}
\value{
  a data frame with three columns
}
\description{
  Useful Information on a Given Dataframe Returns a
  dataframe with the names of each of the columns, the
  class, and number of missing observations
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts
}

