\name{first}
\alias{first}
\title{Peek at the first row of a dataframe. Pretty clever name, no?}
\usage{
  first(dat)
}
\arguments{
  \item{dat}{a dataframe}
}
\value{
  the first row of a dataframe
}
\description{
  Equivalent to head(dat, n = 1)
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts
}
\seealso{
  \code{\link{last}}
}

