\name{lsos}
\alias{lsos}
<<<<<<< HEAD
\title{Useful Version of Objects in Memory}
=======
\title{A more useful summary of objects in the current R environment, including their size in memory}
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\usage{
  lsos(..., n = 10)
}
\arguments{
<<<<<<< HEAD
  \item{\dots}{parameters to pass to \code{lsos}}

  \item{n}{number of rows to return; default is 10}
=======
  \item{...}{paramenters to pass to lsos( )}

  \item{n}{number of rows to return. Default is 10.}
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
}
\description{
  A quick way to see what objects were created in the R
  session and their sizes
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts - Shamelessly ripped from stackoverflow
}
<<<<<<< HEAD
\seealso{
  \code{\link{ls}}
}
=======
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f

