\name{qn}
<<<<<<< HEAD
\alias{.q}
\alias{qn}
\title{Quit the R Session Without Saving Your Workspace}
=======
\alias{qn}
\title{Quit the R session without saving your workspace.}
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\usage{
  qn(...)
}
\arguments{
<<<<<<< HEAD
  \item{\dots}{other parameters for \code{qn()}}
}
\description{
  The save workspace[y/n] prompt is fucking annoying. Avoid
  it with \code{qn()}
=======
  \item{...}{other parameters for qn()}
}
\description{
  The save workspace[y/n] prompt is fucking annoying. Avoid
  it with qn( )
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts - Shamelessly ripped from stackoverflow
}
\seealso{
  \code{\link{q}}
}

