\name{RemoveDupes}
\alias{RemoveDupes}
\title{Remove duplicates from dataset}
\usage{
  RemoveDupes(dat)
}
\arguments{
  \item{dat}{a data.frame}
}
\value{
  a data frame with no duplicate rows
}
\description{
  Removes duplicated rows from a dataframe
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts
}

