\name{SourceNoEcho}
\alias{SourceNoEcho}
\title{Source Files Without Echoing
Small time-saver with an obvious name.}
\usage{
  SourceNoEcho(...)
}
\arguments{
  \item{\dots}{files to source and other options for
  source( )}
}
\value{
  sources files
}
\description{
  Source Files Without Echoing Small time-saver with an
  obvious name.
}
\note{
  Part of a suite of Opower packages
}
\author{
  Yoni Ben-Smellshulam
}
\seealso{
  \code{\link{source}}
}

