\name{tablena}
\alias{tablena}
<<<<<<< HEAD
\title{Easy Way to Include NAs in Table Statements}
=======
\title{An easy way to include NAs in table statements}
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\usage{
  tablena(...)
}
\arguments{
<<<<<<< HEAD
  \item{\dots}{usual parameters passed to \code{table}}
}
\value{
  \code{table()}
}
\description{
  Equivalent to \code{table(..., useNA='ifany')}
=======
  \item{...}{usual parameters passed to table}
}
\value{
  table ( )
}
\description{
  Equivalent to table(..., useNA='ifany')
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
}
\note{
  Part of a suite of Opower packages
}
<<<<<<< HEAD
\examples{
tablena(c(rep(1:5, 5), NA, NA, NA))
}
=======
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\author{
  Erik Shilts
}
\seealso{
  \code{\link{table}}
}

