\name{TopCode}
\alias{TopCode}
\title{Set a maximum value for a numeric vector}
\usage{
  TopCode(vec, level = 100)
}
\arguments{
  \item{vec}{a numeric vector}

  \item{level}{the maximum value. The default is 100.}
}
\value{
  a vector with all values > level set to level
}
\description{
  Top codes a vector by setting a maximum value
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts
}

