\name{BottomCode}
\alias{BottomCode}
\title{Set a minimum value for a numeric vector}
\usage{
  BottomCode(vec, level = 0)
}
\arguments{
  \item{vec}{a numeric vector}

  \item{level}{the minimum value. Default is 0.}
}
\value{
  a vector with all values < level set to level
}
\description{
  Bottom codes a vector by setting a minimum value
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts
}

