\name{classes}
\alias{classes}
\title{Get the classes of columns in a dataframe}
\usage{
  classes(dat)
}
\arguments{
  \item{dat}{a data.frame}
}
\value{
  a vector of class names
}
\description{
  Equivalent to cbind(sapply(dataframe, class))
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts
}

