\name{classes}
\alias{classes}
<<<<<<< HEAD
\title{Classes of Columns in a Dataframe}
=======
\title{Get the classes of columns in a dataframe}
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\usage{
  classes(dat)
}
\arguments{
  \item{dat}{a data.frame}
}
\value{
  a vector of class names
}
\description{
<<<<<<< HEAD
  Equivalent to \code{cbind(sapply(dataframe, class))}
=======
  Equivalent to cbind(sapply(dataframe, class))
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts
}

