\name{concat}
\alias{concat}
\title{Paste Strings Without Having To Set sep = fucking ''}
\usage{
  concat(...)
}
\arguments{
  \item{\dots}{strings to concatenate}
}
\value{
  a concatenated string. meow.
}
\description{
  Otherwise known as concatenate. Seriously, this should be
  obvious.
}
\note{
  Part of a suite of Opower packages
}
\examples{
concat(LETTERS[1:5], letters[1:5])
}
\author{
  Yoni Ben-Smellshulam
}
\seealso{
  \code{\link{paste}}
}

