\name{concat}
\alias{concat}
<<<<<<< HEAD
\title{Paste Strings Without Having To Set sep = fucking ''}
=======
\title{A way to paste strings together without having to set sep = fucking '' over and over again.}
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\usage{
  concat(...)
}
\arguments{
<<<<<<< HEAD
  \item{\dots}{strings to concatenate}
=======
  \item{...}{strings to concatenate}
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
}
\value{
  a concatenated string. meow.
}
\description{
<<<<<<< HEAD
  Otherwise known as concatenate. Seriously, this should be
  obvious.
=======
  Seriously, this should be obvious.
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
}
\note{
  Part of a suite of Opower packages
}
<<<<<<< HEAD
\examples{
concat(LETTERS[1:5], letters[1:5])
}
=======
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\author{
  Yoni Ben-Smellshulam
}
\seealso{
  \code{\link{paste}}
}

