\name{DFInfo}
\alias{DFInfo}
<<<<<<< HEAD
\title{Useful Information on a Given Dataframe}
=======
\title{Get useful information on a given dataframe}
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\usage{
  DFInfo(dat)
}
\arguments{
  \item{dat}{a data frame}
}
\value{
  a data frame with three columns
}
\description{
  Returns a dataframe with the names of each of the
  columns, the class, and number of missing observations
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts
}

