\name{DFInfo}
\alias{DFInfo}
\title{Get useful information on a given dataframe}
\usage{
  DFInfo(dat)
}
\arguments{
  \item{dat}{a data frame}
}
\value{
  a data frame with three columns
}
\description{
  Returns a dataframe with the names of each of the
  columns, the class, and number of missing observations
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts
}

