\name{first}
\alias{first}
<<<<<<< HEAD
\title{First Row of a Dataframe}
=======
\title{Peek at the first row of a dataframe. Pretty clever name, no?}
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\usage{
  first(dat)
}
\arguments{
  \item{dat}{a dataframe}
}
\value{
  the first row of a dataframe
}
\description{
<<<<<<< HEAD
  Equivalent to head(dat, n = 1). Allows you to peek at the
  first line. Pretty clever name, no?
=======
  Equivalent to head(dat, n = 1)
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
}
\note{
  Part of a suite of Opower packages
}
<<<<<<< HEAD
\examples{
temp <- data.frame(var2=1:10, var1=6:15, x=11:20, xy=16:25)
first(temp)
}
=======
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\author{
  Erik Shilts
}
\seealso{
  \code{\link{last}}
}

