\name{lsos}
\alias{lsos}
\title{Useful Version of Objects in Memory}
\usage{
  lsos(..., n = 10)
}
\arguments{
  \item{\dots}{parameters to pass to \code{lsos}}

  \item{n}{number of rows to return; default is 10}
}
\description{
  A quick way to see what objects were created in the R
  session and their sizes
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts - Shamelessly ripped from stackoverflow
}
\seealso{
  \code{\link{ls}}
}

