\name{lsos}
\alias{lsos}
\title{A more useful summary of objects in the current R environment, including their size in memory}
\usage{
  lsos(..., n = 10)
}
\arguments{
  \item{...}{paramenters to pass to lsos( )}

  \item{n}{number of rows to return. Default is 10.}
}
\description{
  A quick way to see what objects were created in the R
  session and their sizes
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts - Shamelessly ripped from stackoverflow
}

