\name{qn}
\alias{qn}
\title{Quit the R session without saving your workspace.}
\usage{
  qn(...)
}
\arguments{
  \item{...}{other parameters for qn()}
}
\description{
  The save workspace[y/n] prompt is fucking annoying. Avoid
  it with qn( )
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts - Shamelessly ripped from stackoverflow
}
\seealso{
  \code{\link{q}}
}

