\name{qn}
\alias{qn}
\title{Quit the R Session Without Saving Your Workspace}
\usage{
  qn(...)
}
\arguments{
  \item{\dots}{other parameters for \code{qn()}}
}
\description{
  The save workspace[y/n] prompt is fucking annoying. Avoid
  it with \code{qn()}
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts - Shamelessly ripped from stackoverflow
}
\seealso{
  \code{\link{q}}
}

