\name{sn}
\alias{sn}
<<<<<<< HEAD
\title{Sorted Names of a Dataframe}
=======
\title{Get the sorted names of a dataframe}
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\usage{
  sn(dat)
}
\arguments{
  \item{dat}{a data.frame}
}
\value{
  a vector of sorted column names
}
\description{
  Equivalent to sort(names(x))
}
\note{
  Part of a suite of Opower packages
}
<<<<<<< HEAD
\examples{
temp <- data.frame(var2=1:10, var1=6:15, x=11:20, xy=16:25)
sn(temp)
}
=======
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\author{
  Erik Shilts
}

