\name{SourceNoEcho}
\alias{SourceNoEcho}
\title{Source Files Without Echoing}
\usage{
  SourceNoEcho(...)
}
\arguments{
  \item{\dots}{files to source and other options for
  source( )}
}
\value{
  sources files
}
\description{
  Small time-saver with an obvious name.
}
\note{
  Part of a suite of Opower packages
}
\author{
  Yoni Ben-Smellshulam
}
\seealso{
  \code{\link{source}}
}

