\name{SourceNoEcho}
\alias{SourceNoEcho}
<<<<<<< HEAD
\title{Source Files Without Echoing}
=======
\title{Source files without echoing them.}
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\usage{
  SourceNoEcho(...)
}
\arguments{
<<<<<<< HEAD
  \item{\dots}{files to source and other options for
  source( )}
=======
  \item{...}{files to source and other options for source(
  )}
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
}
\value{
  sources files
}
\description{
  Small time-saver with an obvious name.
}
\note{
  Part of a suite of Opower packages
}
\author{
  Yoni Ben-Smellshulam
}
\seealso{
  \code{\link{source}}
}

