\name{tablena}
\alias{tablena}
\title{An easy way to include NAs in table statements}
\usage{
  tablena(...)
}
\arguments{
  \item{...}{usual parameters passed to table}
}
\value{
  table ( )
}
\description{
  Equivalent to table(..., useNA='ifany')
}
\note{
  Part of a suite of Opower packages
}
\author{
  Erik Shilts
}
\seealso{
  \code{\link{table}}
}

