\name{TopCode}
\alias{TopCode}
<<<<<<< HEAD
\title{Set a Maximum Value for a Numeric Vector}
=======
\title{Set a maximum value for a numeric vector}
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
\usage{
  TopCode(vec, level = 100)
}
\arguments{
  \item{vec}{a numeric vector}

<<<<<<< HEAD
  \item{level}{the maximum value; default is 100}
}
\value{
  a vector with all values > \code{level} set to
  \code{level}
=======
  \item{level}{the maximum value. The default is 100.}
}
\value{
  a vector with all values > level set to level
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
}
\description{
  Top codes a vector by setting a maximum value
}
\note{
  Part of a suite of Opower packages
}
\author{
<<<<<<< HEAD
  Erik Shilts TopCode(1:10, level=5) TopCode(1:10,
  level=11) TopCode(1:10, level=0)
=======
  Erik Shilts
>>>>>>> 946a93908bb0f9925c07a5e18357b562de0d968f
}

